refactor: renaming file database.py to movie_repository.py to allign with functionallity
